add-recommendation: исправление структуры
