--- src/main/java/ru/yandex/practicum/filmorate/repository/RecommendationStorage.java
@@ -1,9 +1,0 @@
-package ru.yandex.practicum.filmorate.repository;
-
-import ru.yandex.practicum.filmorate.model.Film;
-
-import java.util.Collection;
-
-public interface RecommendationStorage {
-    Collection<Film> getRecommendations(Long userId);
-}
